move handleLogin
